Nombres de comunas de Chile

Vector de texto que contiene el nombre de las 346 comunas de Chile, tal como aparecen en la tabla territorios.

Uso

comunas()

Valor

Vector de nombres de comunas de Chile

Ejemplos

comunas()
#>   [1] "Iquique"              "Alto Hospicio"        "Pozo Almonte"        
#>   [4] "Camiña"               "Colchane"             "Huara"               
#>   [7] "Pica"                 "Antofagasta"          "Mejillones"          
#>  [10] "Sierra Gorda"         "Taltal"               "Calama"              
#>  [13] "Ollagüe"              "San Pedro de Atacama" "Tocopilla"           
#>  [16] "María Elena"          "Copiapó"              "Caldera"             
#>  [19] "Tierra Amarilla"      "Chañaral"             "Diego de Almagro"    
#>  [22] "Vallenar"             "Alto del Carmen"      "Freirina"            
#>  [25] "Huasco"               "La Serena"            "Coquimbo"            
#>  [28] "Andacollo"            "La Higuera"           "Paiguano"            
#>  [31] "Vicuña"               "Illapel"              "Canela"              
#>  [34] "Los Vilos"            "Salamanca"            "Ovalle"              
#>  [37] "Combarbalá"           "Monte Patria"         "Punitaqui"           
#>  [40] "Río Hurtado"          "Valparaíso"           "Casablanca"          
#>  [43] "Concón"               "Juan Fernández"       "Puchuncaví"          
#>  [46] "Quintero"             "Viña del Mar"         "Isla de Pascua"      
#>  [49] "Los Andes"            "Calle Larga"          "Rinconada"           
#>  [52] "San Esteban"          "La Ligua"             "Cabildo"             
#>  [55] "Papudo"               "Petorca"              "Zapallar"            
#>  [58] "Quillota"             "Calera"               "Hijuelas"            
#>  [61] "La Cruz"              "Nogales"              "San Antonio"         
#>  [64] "Algarrobo"            "Cartagena"            "El Quisco"           
#>  [67] "El Tabo"              "Santo Domingo"        "San Felipe"          
#>  [70] "Catemu"               "Llaillay"             "Panquehue"           
#>  [73] "Putaendo"             "Santa María"          "Quilpué"             
#>  [76] "Limache"              "Olmué"                "Villa Alemana"       
#>  [79] "Rancagua"             "Codegua"              "Coinco"              
#>  [82] "Coltauco"             "Doñihue"              "Graneros"            
#>  [85] "Las Cabras"           "Machalí"              "Malloa"              
#>  [88] "Mostazal"             "Olivar"               "Peumo"               
#>  [91] "Pichidegua"           "Quinta de Tilcoco"    "Rengo"               
#>  [94] "Requínoa"             "San Vicente"          "Pichilemu"           
#>  [97] "La Estrella"          "Litueche"             "Marchihue"           
#> [100] "Navidad"              "Paredones"            "San Fernando"        
#> [103] "Chépica"              "Chimbarongo"          "Lolol"               
#> [106] "Nancagua"             "Palmilla"             "Peralillo"           
#> [109] "Placilla"             "Pumanque"             "Santa Cruz"          
#> [112] "Talca"                "Constitución"         "Curepto"             
#> [115] "Empedrado"            "Maule"                "Pelarco"             
#> [118] "Pencahue"             "Río Claro"            "San Clemente"        
#> [121] "San Rafael"           "Cauquenes"            "Chanco"              
#> [124] "Pelluhue"             "Curicó"               "Hualañé"             
#> [127] "Licantén"             "Molina"               "Rauco"               
#> [130] "Romeral"              "Sagrada Familia"      "Teno"                
#> [133] "Vichuquén"            "Linares"              "Colbún"              
#> [136] "Longaví"              "Parral"               "Retiro"              
#> [139] "San Javier"           "Villa Alegre"         "Yerbas Buenas"       
#> [142] "Concepción"           "Coronel"              "Chiguayante"         
#> [145] "Florida"              "Hualqui"              "Lota"                
#> [148] "Penco"                "San Pedro de la Paz"  "Santa Juana"         
#> [151] "Talcahuano"           "Tomé"                 "Hualpén"             
#> [154] "Lebu"                 "Arauco"               "Cañete"              
#> [157] "Contulmo"             "Curanilahue"          "Los Alamos"          
#> [160] "Tirúa"                "Los Angeles"          "Antuco"              
#> [163] "Cabrero"              "Laja"                 "Mulchén"             
#> [166] "Nacimiento"           "Negrete"              "Quilaco"             
#> [169] "Quilleco"             "San Rosendo"          "Santa Bárbara"       
#> [172] "Tucapel"              "Yumbel"               "Alto Biobío"         
#> [175] "Temuco"               "Carahue"              "Cunco"               
#> [178] "Curarrehue"           "Freire"               "Galvarino"           
#> [181] "Gorbea"               "Lautaro"              "Loncoche"            
#> [184] "Melipeuco"            "Nueva Imperial"       "Padre Las Casas"     
#> [187] "Perquenco"            "Pitrufquén"           "Pucón"               
#> [190] "Saavedra"             "Teodoro Schmidt"      "Toltén"              
#> [193] "Vilcún"               "Villarrica"           "Cholchol"            
#> [196] "Angol"                "Collipulli"           "Curacautín"          
#> [199] "Ercilla"              "Lonquimay"            "Los Sauces"          
#> [202] "Lumaco"               "Purén"                "Renaico"             
#> [205] "Traiguén"             "Victoria"             "Puerto Montt"        
#> [208] "Calbuco"              "Cochamó"              "Fresia"              
#> [211] "Frutillar"            "Los Muermos"          "Llanquihue"          
#> [214] "Maullín"              "Puerto Varas"         "Castro"              
#> [217] "Ancud"                "Chonchi"              "Curaco de Vélez"     
#> [220] "Dalcahue"             "Puqueldón"            "Queilén"             
#> [223] "Quellón"              "Quemchi"              "Quinchao"            
#> [226] "Osorno"               "Puerto Octay"         "Purranque"           
#> [229] "Puyehue"              "Río Negro"            "San Juan de la Costa"
#> [232] "San Pablo"            "Chaitén"              "Futaleufú"           
#> [235] "Hualaihué"            "Palena"               "Coyhaique"           
#> [238] "Lago Verde"           "Aysén"                "Cisnes"              
#> [241] "Guaitecas"            "Cochrane"             "O'Higgins"           
#> [244] "Tortel"               "Chile Chico"          "Río Ibáñez"          
#> [247] "Punta Arenas"         "Laguna Blanca"        "Río Verde"           
#> [250] "San Gregorio"         "Cabo de Hornos"       "Antártica"           
#> [253] "Porvenir"             "Primavera"            "Timaukel"            
#> [256] "Natales"              "Torres del Paine"     "Santiago"            
#> [259] "Cerrillos"            "Cerro Navia"          "Conchalí"            
#> [262] "El Bosque"            "Estación Central"     "Huechuraba"          
#> [265] "Independencia"        "La Cisterna"          "La Florida"          
#> [268] "La Granja"            "La Pintana"           "La Reina"            
#> [271] "Las Condes"           "Lo Barnechea"         "Lo Espejo"           
#> [274] "Lo Prado"             "Macul"                "Maipú"               
#> [277] "Ñuñoa"                "Pedro Aguirre Cerda"  "Peñalolén"           
#> [280] "Providencia"          "Pudahuel"             "Quilicura"           
#> [283] "Quinta Normal"        "Recoleta"             "Renca"               
#> [286] "San Joaquín"          "San Miguel"           "San Ramón"           
#> [289] "Vitacura"             "Puente Alto"          "Pirque"              
#> [292] "San José de Maipo"    "Colina"               "Lampa"               
#> [295] "Tiltil"               "San Bernardo"         "Buin"                
#> [298] "Calera de Tango"      "Paine"                "Melipilla"           
#> [301] "Alhué"                "Curacaví"             "María Pinto"         
#> [304] "San Pedro"            "Talagante"            "El Monte"            
#> [307] "Isla de Maipo"        "Padre Hurtado"        "Peñaflor"            
#> [310] "Valdivia"             "Corral"               "Lanco"               
#> [313] "Los Lagos"            "Máfil"                "Mariquina"           
#> [316] "Paillaco"             "Panguipulli"          "La Unión"            
#> [319] "Futrono"              "Lago Ranco"           "Río Bueno"           
#> [322] "Arica"                "Camarones"            "Putre"               
#> [325] "General Lagos"        "Chillán"              "Bulnes"              
#> [328] "Chillán Viejo"        "El Carmen"            "Pemuco"              
#> [331] "Pinto"                "Quillón"              "San Ignacio"         
#> [334] "Yungay"               "Quirihue"             "Cobquecura"          
#> [337] "Coelemu"              "Ninhue"               "Portezuelo"          
#> [340] "Ranquil"              "Treguaco"             "San Carlos"          
#> [343] "Coihueco"             "Ñiquén"               "San Fabián"          
#> [346] "San Nicolás"