Tabla de localidades de Chile

Tabla de datos que contiene todas las localidades de Chile, asentamientos o áreas subcomunales que no son entidades administrativas. Según la definición del INE, una localidad es un "Área geográfica con nombre propio de conocimiento generalizado por la población. Puede estar poblada o no, aunque debe contener, a lo menos, una vivienda susceptible de ser habitada. Generalmente, se circunscribe dentro de un distrito censal; no obstante, puede rebasar a éste.” Esta tabla de localidades proviene del estudio de Identificación de localidades en condición de aislamiento, realizado por la Subsecretaría de Desarrollo Regional y Administrativo y publicado el año 2021.

Uso

localidades

Formato

Un data frame con 29.256 filas y 6 columnas:

codigo_comuna

Código único territorial de las comunas de Chile

localidad

Nombre de la localidad

tipo

Tipo de localidad

subtipo

Subtipo de la localidad

habitantes

Habitantes de la localidad

viviendas

Viviendas de la localidad

Fuente

https://ide.subdere.gov.cl/estudio-actualizacion-de-base-censal-identificacion-de-localidades-en-condicion-de-aislamiento/