Tabla de comunas, provincias y regiones de Chile

Tabla de datos que contiene todas las comunas del país con sus nombres oficiales, sus códigos únicos comunales, y lo mismo para las provincias y regiones del país. Al ser una tabla de las comunas de Chile, tiene 346 filas. Estos datos corresponden al Ministerio del Interior de Chile, publicados en el Diario Oficial el 21 de septiembre de 2018, con leves modificaciones.

Uso

territorios

Formato

A data frame with 346 rows and 6 columns:

codigo_region

Código único territorial de las regiones de Chile (número entre 1 y 16)

nombre_region

Nombre de las regiones de Chile

codigo_provincia

Código único territorial de las provincias de Chile

nombre_provincia

Nombre de las provincias de Chile

codigo_comuna

Código único territorial de las comunas de Chile

nombre_comuna

Nombre de las comunas de Chile

Fuente

https://www.subdere.gov.cl/documentacion/códigos-únicos-territoriales-actualizados-al-06-de-septiembre-2018